move node too
