+++ slk/chain/chain.py
@@ -22,9 +22,9 @@ from xrpl.models import (
 from xrpl.models.transactions.transaction import Transaction
 from xrpl.utils import drops_to_xrp
 
+from slk.chain.node import Node
 from slk.common import Account
 from slk.config_file import ConfigFile
-from slk.node import Node
 
 
 class KeyManager:
