+++ slk/chain/testnet.py
@@ -10,8 +10,8 @@ from typing import Any, Dict, Generator, List, Optional, Set, Union
 from xrpl.models import FederatorInfo
 
 from slk.chain.chain import Chain
+from slk.chain.node import Node
 from slk.config_file import ConfigFile
-from slk.node import Node
 
 
 class Sidechain(Chain):
